requirements.txt
numpy==1.15.4
opencv-contrib-python==3.4.4.19
opencv-python==3.4.4.19
Pillow==5.3.0
PyAutoGUI==0.9.38
PyMsgBox==1.0.6
pypi==2.1
PyScreeze==0.1.18
PyTweening==1.0.3
pytz==2018.7
